requirements.txt
amqp==5.1.1
asgiref==3.7.2
celery==5.2.7
Django==4.2.1
django-admin-sortable2==2.1.8
django-ckeditor==6.5.1
django-cors-headers==4.0.0
django-gfklookupwidget==1.0.9
django-import-export==3.2.0
djangorestframework==3.14.0
drf-spectacular==0.26.2
psycopg2-binary==2.9.6
redis==4.5.5
requests==2.31.0
pre-commit~=3.4.0
Pillow==10.0.0
isort ~= 5.1.0
eyed3 == 0.9.7